requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
google-generativeai==0.8.6
pydantic==2.10.6

